requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
uvicorn
streamlit
langchain-groq
chromadb
langchain-chroma
ipykernel
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
sentence_transformers
langchain_huggingface
duckdb
pandas
openai
langchain-groq
duckduckgo-search
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
email-validator==2.0
validators==0.28.1
youtube_transcript_api
huggingface_hub
gradio


